Bug Fix on addition, added normalise proc
I have been meaning to refactor the addition operation for a while because it was a bit ugly. However, I found during testing that in some instances incorrect results were being produced due to incorrect normalisation logic (when dealing with different exponents). As a result I have reverted to the logic from python's decimal for now so that I can be sure the results are accurate, and I will worry about optimisation/handling special cases later