Update toString procs
While there is little performance difference between the existing string conversion methods and the cpython implementation, that version is a lot cleaner to look at. For that reason I am converting both scientific and engineering string procedures to the cpython implementation. Also added a little more structure to context but I am still experimenting with that