Field redesign (strings -> bigints)
The decision to try to use a coefficient stored as string made simple arithmetic operations take roughly 200% of the time of using big-ints. I will re-implement the Decimal type without as much use of strings to start with for an immediate performance improvement.

Ideally I can remove the use of strings as much as possible as conversions are where a lot of time is being spent. However, that will still be orders of magnitude slower than efficient C implementations such as decNumber. If you are seeking an efficient Decimal type, it would be better to create a wrapper for an existing C implementation.

This library is only useful as a learning tool at the moment due to the weak performance, but I should at least attempt to make it as efficient as possible if I can without a major redesign.